refactor: migrate type imports from types to runtime and remove unused types module
