+++ custom_components/echonet_lite/binary_sensor.py
@@ -20,7 +20,7 @@ from .entity import (
     setup_echonet_lite_platform,
 )
 from .prop import BinaryProp
-from .types import EchonetLiteConfigEntry
+from .runtime import EchonetLiteConfigEntry
 
 PARALLEL_UPDATES = 0
