+++ custom_components/echonet_lite/select.py
@@ -35,7 +35,7 @@ from .entity import (
     setup_echonet_lite_platform,
 )
 from .prop import EnumProp
-from .types import EchonetLiteConfigEntry
+from .runtime import EchonetLiteConfigEntry
 
 PARALLEL_UPDATES = 1
