+++ custom_components/echonet_lite/sensor.py
@@ -26,7 +26,7 @@ from .entity import (
     setup_echonet_lite_platform,
 )
 from .prop import EnumProp, NumericProp
-from .types import EchonetLiteConfigEntry
+from .runtime import EchonetLiteConfigEntry
 
 PARALLEL_UPDATES = 0
