+++ custom_components/echonet_lite/switch.py
@@ -17,7 +17,7 @@ from .entity import (
     setup_echonet_lite_platform,
 )
 from .prop import BinaryProp
-from .types import EchonetLiteConfigEntry
+from .runtime import EchonetLiteConfigEntry
 
 PARALLEL_UPDATES = 1
